refactor: remove CodeEditorWrapper, move CodeEditor to Shared/Components, update instances of CodeEditor
