+++ src/Shared/Components/CodeEditor/Commands/findAndReplace.ts
