+++ src/Shared/Components/CodeEditor/Extensions/DiffMinimap.tsx
