+++ src/Shared/Components/CodeEditor/Extensions/findAndReplace.tsx
