+++ src/Shared/Components/CodeEditor/Extensions/readOnlyTooltip.ts
