+++ src/Shared/Components/CodeEditor/codeEditor.scss
