--- src/Shared/Components/CodeEditorWrapper/types.ts
@@ -1,52 +1,0 @@
-/*
- * Copyright (c) 2024. Devtron Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { CodeEditorProps } from '@Common/CodeMirror'
-
-export type CodeEditorWrapperProps<DiffView extends boolean> = Pick<
-    CodeEditorProps<DiffView>,
-    | 'mode'
-    | 'tabSize'
-    | 'readOnly'
-    | 'placeholder'
-    | 'noParsing'
-    | 'loading'
-    | 'customLoader'
-    | 'cleanData'
-    | 'disableSearch'
-    | 'children'
-> & {
-    diffView?: DiffView
-    /**
-     * @deprecated
-     * Monaco editor has been removed. \
-     * This prop is no longer required and will be removed in future releases.
-     */
-    codeEditorProps?: any
-    codeMirrorProps: Omit<
-        CodeEditorProps<DiffView>,
-        | 'mode'
-        | 'tabSize'
-        | 'readOnly'
-        | 'placeholder'
-        | 'noParsing'
-        | 'loading'
-        | 'customLoader'
-        | 'cleanData'
-        | 'disableSearch'
-        | 'children'
-    >
-}
