+++ src/Shared/Components/index.ts
@@ -30,7 +30,7 @@ export * from './ButtonWithLoader'
 export * from './ButtonWithSelector'
 export * from './CICDHistory'
 export * from './CMCS'
-export * from './CodeEditorWrapper'
+export * from './CodeEditor'
 export * from './Collapse'
 export * from './CollapsibleList'
 export * from './CommitChipCell'
